refactor(front): corrigindo estilos de machine e machine-list; criando componente machine-card
