+++ Gerenciador de Contratos/src/pages/machine-list/index.tsx
@@ -1,92 +1,21 @@
 import { Machine } from "@/interfaces/machine";
 import Layout from "@/layouts/default";
 import {
-  loadMachineImage,
   loadUserMachines,
 } from "@/services/api/machine/machine";
 import { useEffect, useState } from "react";
 import {
   Card,
   CardContent,
-  CardDescription,
   CardHeader,
 } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
 import "@/components/machine-list/machine-list.css";
 import { Machine as Maquina } from "@/interfaces/machine";
-import { useNavigate } from "react-router-dom";
 import { Button } from "@/components/ui/button";
 import MachineFilter from "@/components/machine-filter";
+import { MachineCard } from "@/components/machine-card";
 
 export const MachineList = () => {
-  const MachineCard: React.FC<{ machine: Maquina }> = ({ machine }) => {
-    const [image, setImage] = useState("");
-    const [loadingImage, setLoadingImage] = useState(true);
-    const [error, setError] = useState(false);
-    const navigate = useNavigate();
-
-    const fetchMachineImage = async (machineId: string) => {
-      try {
-        console.log("machineId: ", machineId);
-        const response = await loadMachineImage(machineId);
-        const imageUrl = response;
-        setImage(imageUrl);
-        setLoadingImage(false);
-      } catch (error) {
-        console.error(error);
-        setError(true);
-      } finally {
-        setLoadingImage(false);
-      }
-    };
-
-    useEffect(() => {
-      if (machine) {
-        fetchMachineImage(machine.idmaquina);
-      }
-    }, [machine]);
-
-    return (
-      <Card
-        className={`machine-list-card ${
-          machines.length === 1 ? "single-item" : ""
-        } hover:cursor-pointer`}
-        onClick={() => {
-          navigate(`/machine-details/${machine?.idpublico}`);
-        }}
-      >
-        <CardContent>
-          <div className="machine-list-image">
-            {loadingImage ? (
-              <div>
-                <Skeleton className="h-[30vh] w-[30vw] rounded-xl" />
-              </div>
-            ) : error ? (
-              <div className="image-placeholder">Erro ao carregar imagem</div>
-            ) : image ? (
-              <img
-                src={image}
-                className="p-0 w-full"
-                alt={`Imagem de ${machine?.nome}`}
-              />
-            ) : (
-              <div className="image-placeholder">Imagem indisponível</div>
-            )}
-          </div>
-          <h1>{machine?.nome}</h1>
-          <CardDescription>
-            <p className="machine-list-card-description">
-              Preço: <strong>R$ {machine?.valoraluguel}</strong>
-            </p>
-            <p className="machine-list-card-description">
-              Categoria: {machine?.categoria}
-            </p>
-          </CardDescription>
-        </CardContent>
-      </Card>
-    );
-  };
-
   const [machines, setMachines] = useState<Machine[]>([]);
   const [filter, setFilter] = useState("");
 
@@ -149,8 +78,7 @@ export const MachineList = () => {
             filteredMachines.map((machine: Maquina) => (
               <div
                 key={machine.idmaquina}
-                className={`machine-list-card ${filteredMachines.length === 1 ? 'single-item' : ''}`}
-                style={{ marginTop: "4vh", width: "90%", height: "768px" }}
+                style={{ width: "90%", height: "500px", padding: '0' }}
               >
                 <MachineCard machine={machine} />
               </div>
